Initial refactor of the visualisers
